# docs/openapi/swagger.yaml
openapi: 3.0.3
info:
  title: Azure DevOps to GitHub Migration Operator API
  description: |
    API for the Azure DevOps to GitHub Migration Operator, which facilitates the migration of resources from Azure DevOps to GitHub.
    
    This API provides endpoints for:
    - Discovering Azure DevOps resources (organizations, projects, repositories, work items, pipelines)
    - Creating and managing migrations
    - Monitoring migration progress
    - Converting Azure DevOps pipelines to GitHub Actions workflows
  version: 1.0.0
  contact:
    name: ADO to Git Migration Team
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /api/v1
    description: API base path
...
paths:
  /discovery/organizations:
    get:
      tags:
        - Discovery
      summary: Discover Azure DevOps organizations
      description: Retrieves a list of Azure DevOps organizations accessible to the authenticated user
      parameters:
        - name: client_id
          in: query
          description: Azure Service Principal Client ID
          required: true
          schema:
            type: string
            format: uuid
        - name: client_secret
          in: query
          description: Azure Service Principal Client Secret
          required: true
          schema:
            type: string
        - name: tenant_id
          in: query
          description: Azure Tenant ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Organization'
                  count:
                    type: integer
                    description: Total number of organizations
        '400':
          description: Bad request - missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /discovery/projects:
    get:
      tags:
        - Discovery
      summary: Discover Azure DevOps projects
      description: Retrieves a list of projects in an Azure DevOps organization
      parameters:
        - name: organization
          in: query
          description: Azure DevOps organization name
          required: true
          schema:
            type: string
        - name: client_id
          in: query
          description: Azure Service Principal Client ID
          required: true
          schema:
            type: string
            format: uuid
        - name: client_secret
          in: query
          description: Azure Service Principal Client Secret
          required: true
          schema:
            type: string
        - name: tenant_id
          in: query
          description: Azure Tenant ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  count:
                    type: integer
                    description: Total number of projects
                  organization:
                    type: string
                    description: Organization name
        '400':
          description: Bad request - missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /discovery/repositories:
    get:
      tags:
        - Discovery
      summary: Discover Azure DevOps repositories
      description: Retrieves a list of repositories in an Azure DevOps project
      parameters:
        - name: organization
          in: query
          description: Azure DevOps organization name
          required: true
          schema:
            type: string
        - name: project
          in: query
          description: Azure DevOps project name
          required: true
          schema:
            type: string
        - name: client_id
          in: query
          description: Azure Service Principal Client ID
          required: true
          schema:
            type: string
            format: uuid
        - name: client_secret
          in: query
          description: Azure Service Principal Client Secret
          required: true
          schema:
            type: string
        - name: tenant_id
          in: query
          description: Azure Tenant ID
          required: true
          schema:
            type: string
            format: uuid
        - name: include_empty
          in: query
          description: Include empty repositories
          required: false
          schema:
            type: boolean
            default: false
        - name: include_disabled
          in: query
          description: Include disabled repositories
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  repositories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Repository'
                  count:
                    type: integer
                    description: Total number of repositories
                  organization:
                    type: string
                    description: Organization name
                  project:
                    type: string
                    description: Project name
        '400':
          description: Bad request - missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /discovery/workitems:
    get:
      tags:
        - Discovery
      summary: Discover Azure DevOps work items
      description: Retrieves a list of work items in an Azure DevOps project
      parameters:
        - name: organization
          in: query
          description: Azure DevOps organization name
          required: true
          schema:
            type: string
        - name: project
          in: query
          description: Azure DevOps project name
          required: true
          schema:
            type: string
        - name: client_id
          in: query
          description: Azure Service Principal Client ID
          required: true
          schema:
            type: string
            format: uuid
        - name: client_secret
          in: query
          description: Azure Service Principal Client Secret
          required: true
          schema:
            type: string
        - name: tenant_id
          in: query
          description: Azure Tenant ID
          required: true
          schema:
            type: string
            format: uuid
        - name: types
          in: query
          description: Comma-separated work item types
          required: false
          schema:
            type: string
        - name: states
          in: query
          description: Comma-separated work item states
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of items to return
          required: false
          schema:
            type: integer
            default: 1000
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  workItems:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkItem'
                  count:
                    type: integer
                    description: Total number of work items
                  organization:
                    type: string
                    description: Organization name
                  project:
                    type: string
                    description: Project name
        '400':
          description: Bad request - missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /migrations:
    get:
      tags:
        - Migrations
      summary: List migrations
      description: Retrieves a list of all migration jobs
      parameters:
        - name: namespace
          in: query
          description: Kubernetes namespace
          required: false
          schema:
            type: string
            default: default
        - name: phase
          in: query
          description: Filter by migration phase
          required: false
          schema:
            type: string
            enum: [Pending, Validating, Running, Completed, Failed, Cancelled, Paused]
        - name: type
          in: query
          description: Filter by migration type
          required: false
          schema:
            type: string
            enum: [repository, workitems, pipelines, all]
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  migrations:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdoToGitMigration'
                  count:
                    type: integer
                    description: Total number of migrations
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Migrations
      summary: Create migration
      description: Creates a new migration job
      requestBody:
        description: Migration job configuration
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdoToGitMigration'
      responses:
        '201':
          description: Migration created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdoToGitMigration'
        '400':
          description: Bad request - invalid configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /migrations/{id}:
    get:
      tags:
        - Migrations
      summary: Get migration
      description: Retrieves details of a specific migration job
      parameters:
        - name: id
          in: path
          description: Migration job name
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          description: Kubernetes namespace
          required: false
          schema:
            type: string
            default: default
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdoToGitMigration'
        '404':
          description: Migration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Migrations
      summary: Delete migration
      description: Deletes a specific migration job
      parameters:
        - name: id
          in: path
          description: Migration job name
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          description: Kubernetes namespace
          required: false
          schema:
            type: string
            default: default
      responses:
        '204':
          description: Migration deleted successfully
        '404':
          description: Migration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /migrations/{id}/status:
    get:
      tags:
        - Migrations
      summary: Get migration status
      description: Retrieves the status of a specific migration job
      parameters:
        - name: id
          in: path
          description: Migration job name
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          description: Kubernetes namespace
          required: false
          schema:
            type: string
            default: default
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdoToGitMigrationStatus'
        '404':
          description: Migration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
        status:
          type: integer
          description: HTTP status code
    Organization:
      type: object
      properties:
        id:
          type: string
          description: Organization ID
        name:
          type: string
          description: Organization name
        url:
          type: string
          format: uri
          description: Organization URL
        description:
          type: string
          description: Organization description
        region:
          type: string
          description: Azure region
        owner:
          type: string
          description: Organization owner
        createdDate:
          type: string
          format: date-time
          description: Creation date
        lastModifiedDate:
          type: string
          format: date-time
          description: Last modification date
      required:
        - id
        - name
        - url
    Project:
      type: object
      properties:
        id:
          type: string
          description: Project ID
        name:
          type: string
          description: Project name
        description:
          type: string
          description: Project description
        url:
          type: string
          format: uri
          description: Project URL
        state:
          type: string
          description: Project state
          enum: [wellFormed, createPending, deleting, new, all]
        visibility:
          type: string
          description: Project visibility
          enum: [private, public]
        revision:
          type: integer
          format: int64
          description: Project revision
        lastUpdateTime:
          type: string
          format: date-time
          description: Last update time
        capabilities:
          type: object
          additionalProperties: true
          description: Project capabilities
      required:
        - id
        - name
        - url
        - state
        - visibility
        - revision
    Repository:
      type: object
      properties:
        id:
          type: string
          description: Repository ID
        name:
          type: string
          description: Repository name
        url:
          type: string
          format: uri
          description: Repository URL
        defaultBranch:
          type: string
          description: Default branch
        size:
          type: integer
          format: int64
          description: Repository size in bytes
        isEmpty:
          type: boolean
          description: Whether the repository is empty
        isDisabled:
          type: boolean
          description: Whether the repository is disabled
        projectId:
          type: string
          description: Parent project ID
        projectName:
          type: string
          description: Parent project name
        remoteUrl:
          type: string
          format: uri
          description: Remote URL
        sshUrl:
          type: string
          description: SSH URL
        webUrl:
          type: string
          format: uri
          description: Web URL
        branchCount:
          type: integer
          description: Number of branches
        tagCount:
          type: integer
          description: Number of tags
        pullRequestCount:
          type: integer
          description: Number of pull requests
        lastCommit:
          $ref: '#/components/schemas/Commit'
      required:
        - id
        - name
        - url
        - defaultBranch
        - size
        - isEmpty
        - isDisabled
        - projectId
        - projectName
        - remoteUrl
        - webUrl
    Commit:
      type: object
      properties:
        id:
          type: string
          description: Commit ID
        author:
          type: string
          description: Author
        committer:
          type: string
          description: Committer
        comment:
          type: string
          description: Commit message
        date:
          type: string
          format: date-time
          description: Commit date
        url:
          type: string
          format: uri
          description: Commit URL
      required:
        - id
        - author
        - committer
        - comment
        - date
        - url
    WorkItem:
      type: object
      properties:
        id:
          type: integer
          description: Work item ID
        type:
          type: string
          description: Work item type
        title:
          type: string
          description: Work item title
        state:
          type: string
          description: Work item state
        assignedTo:
          type: string
          description: Assigned user
        createdDate:
          type: string
          format: date-time
          description: Creation date
        changedDate:
          type: string
          format: date-time
          description: Last change date
        areaPath:
          type: string
          description: Area path
        iterationPath:
          type: string
          description: Iteration path
        tags:
          type: array
          items:
            type: string
          description: Work item tags
        priority:
          type: string
          description: Priority
        severity:
          type: string
          description: Severity
        url:
          type: string
          format: uri
          description: Work item URL
        projectId:
          type: string
          description: Parent project ID
        projectName:
          type: string
          description: Parent project name
      required:
        - id
        - type
        - title
        - state
        - createdDate
        - changedDate
        - url
        - projectId
        - projectName
    AdoToGitMigration:
      type: object
      properties:
        apiVersion:
          type: string
          default: migration.ado-to-git-migration.io/v1
        kind:
          type: string
          default: AdoToGitMigration
        metadata:
          type: object
          properties:
            name:
              type: string
              description: Migration name
            namespace:
              type: string
              description: Kubernetes namespace
            creationTimestamp:
              type: string
              format: date-time
              description: Creation timestamp
        spec:
          $ref: '#/components/schemas/AdoToGitMigrationSpec'
        status:
          $ref: '#/components/schemas/AdoToGitMigrationStatus'
      required:
        - apiVersion
        - kind
        - metadata
        - spec
    AdoToGitMigrationSpec:
      type: object
      properties:
        type:
          type: string
          description: Migration type
          enum: [repository, workitems, pipelines, all]
        source:
          $ref: '#/components/schemas/AdoSourceConfig'
        target:
          $ref: '#/components/schemas/GitHubTargetConfig'
        settings:
          $ref: '#/components/schemas/MigrationSettings'
        resources:
          type: array
          items:
            $ref: '#/components/schemas/MigrationResource'
          description: Resources to migrate
        validationRules:
          $ref: '#/components/schemas/ValidationRules'
      required:
        - type
        - source
        - target
        - settings
        - resources
...
    ValidationRules:
      type: object
      properties:
        skipValidation:
          type: boolean
          description: Skip validation
          default: false
        requiredPermissions:
          type: array
          items:
            type: string
          description: Required permissions
        preMigrationChecks:
          type: array
          items:
            type: string
          description: Pre-migration checks
    AdoToGitMigrationStatus:
      type: object
      properties:
        phase:
          type: string
          description: Migration phase
          enum: [Pending, Validating, Running, Completed, Failed, Cancelled, Paused]
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/Condition'
          description: Status conditions
        startTime:
          type: string
          format: date-time
          description: Start time
        completionTime:
          type: string
          format: date-time
          description: Completion time
        progress:
          $ref: '#/components/schemas/MigrationProgress'
        resourceStatuses:
          type: array
          items:
            $ref: '#/components/schemas/ResourceMigrationStatus'
          description: Resource migration statuses
        validationResults:
          $ref: '#/components/schemas/ValidationResults'
        errorMessage:
          type: string
          description: Error message
        warnings:
          type: array
          items:
            type: string
          description: Warning messages
        statistics:
          $ref: '#/components/schemas/MigrationStatistics'
      required:
        - phase
        - progress
    Condition:
      type: object
      properties:
        type:
          type: string
          description: Condition type
        status:
          type: string
          description: Condition status
          enum: ["True", "False", "Unknown"]
        lastTransitionTime:
          type: string
          format: date-time
          description: Last transition time
        reason:
          type: string
          description: Reason for the condition
        message:
          type: string
          description: Human-readable message
      required:
        - type
        - status
        - lastTransitionTime
        - reason
        - message
...